g1

+++ pybricks/pyBricks1/main.py
@@ -92,7 +92,7 @@ async def main():
     await WheelController.wheel_right_turn_with_angle(float(30))
     await WheelController.move_wheels_forward_in_straight_line(float(300))
     await WheelController.wheel_left_turn()
-    await WheelController.move_wheels_forward_in_straight_line(float(30))
+    await WheelController.move_wheels_forward_in_straight_line(float(60))
     await WheelController.move_wheels_towards_element_then_stop_at_marker()
     print("DONE!")
